% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EpiModel.hpc-package.r
\docType{package}
\name{EpiModel.hpc-package}
\alias{EpiModel.hpc}
\alias{EpiModel.hpc-package}
\title{EpiModel Extensions for High-Performance Computing}
\description{
\tabular{ll}{
   Package: \tab EpiModel.hpc\cr
   Type: \tab Package\cr
   Version: \tab 1.0\cr
   Date: \tab 2015-03-23\cr
   License: \tab GPL (>= 2)\cr
   LazyLoad: \tab yes\cr
}
}
\details{
The text here
}
\references{
The main website is at \url{http://epimodel.org/},
            and the source code is at \url{http://github.com/statnet/EpiModel.hpc}.
            Bug reports and feature requests may be filed there.
}
\keyword{package}

