% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModelHPC-package.r
\docType{package}
\name{EpiModelHPC-package}
\alias{EpiModelHPC-package}
\alias{EpiModelHPC}
\title{EpiModel Extensions for High-Performance Computing}
\description{
Extension package to EpiModel to run large-scale stochastic network models on modern high-performance computing systems. Provides scenario-based batch simulation, helpers for building 'slurmworkflow' job workflows and 'swfcalib' calibrations, cluster configuration presets, and shell tooling for detecting and recovering degenerate SLURM tasks.
}
\details{
EpiModel provides tools for the mathematical modeling of infectious diseases.
Supported model classes include stochastic network models, which rely on the
statistical framework of exponential-family random graph models (ERGMs) that
evolve over time. This allows for modeling of disease-related contacts with
duration, such as ongoing sexual partnerships.

The level of statistical complexity of these models, based in Markov-chain
Monte Carlo (MCMC) simulation, results in computationally intensive
simulation processes. The goal of EpiModelHPC is to provide a standardized
framework for extending EpiModel to run on modern high-performance computing
(HPC) systems.
}
\references{
The main website for EpiModel is at \url{https://www.epimodel.org/}.
The source code for this extension package is hosted on GitHub
at \url{https://github.com/EpiModel/EpiModelHPC}. Bug reports and
feature requests may be filed there.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.epimodel.org/}
  \item \url{https://epimodel.github.io/EpiModelHPC/}
  \item Report bugs at \url{https://github.com/EpiModel/EpiModelHPC/issues}
}

}
\author{
\strong{Maintainer}: Samuel Jenness \email{samuel.m.jenness@emory.edu}

Authors:
\itemize{
  \item Samuel Jenness \email{samuel.m.jenness@emory.edu}
}

}
\keyword{package}
