% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merge_simfiles.R
\name{get_epi}
\alias{get_epi}
\title{Extract Epidemiological Output from Saved Simulation Files and Resave.}
\usage{
get_epi(simno = NA, indir = "data/", outdir = "data/", vars = NULL,
  suffix = "lim", compress = "xz")
}
\arguments{
\item{simno}{Simulation number to process.}

\item{indir}{File directory relative to working directory where simulation
files are stored.}

\item{outdir}{File directory relative to working directory where simulation
files should be saved.}

\item{vars}{Argument passed to \code{\link{merge_simfiles}}.}

\item{suffix}{Suffix to add to file name before saving the extracted files.}

\item{compress}{Argument passed to \code{\link{save}}.}
}
\description{
Loads a series of simulation files processed with \code{process_simfiles},
             extracts epidemiological output for smaller file sizes, and then
             saves the output.
}

