% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netsim_scenarios.R
\name{get_scenarios_tibble_infos}
\alias{get_scenarios_tibble_infos}
\title{Helper function to access the infos on merged scenarios \code{data.frame}}
\usage{
get_scenarios_tibble_infos(scenario_dir)
}
\arguments{
\item{scenario_dir}{the directory where \code{merge_netsim_scenarios_tibble} saved
the merged tibbles.}
}
\value{
a \code{tibble} with two columns: \code{file_path} - the full path of
the scenario tibble file and \code{scenario_name} the associated scenario name.
}
\description{
This function returns the list of scenario tibble files and the corresponding
scenario name present in a given directory. It is meant to
be used after \code{merge_netsim_scenarios_tibble} or
\code{step_tmpl_merge_netsim_scenarios_tibble}.
}
