% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initialize_cp.R
\name{initialize_cp}
\alias{initialize_cp}
\title{Initializes Network Model after Checkpointing}
\usage{
initialize_cp(x, param, init, control, s)
}
\arguments{
\item{x}{An \code{EpiModel} object of class \code{\link{netest}}.}

\item{param}{An \code{EpiModel} object of class \code{\link{param.net}}.}

\item{init}{An \code{EpiModel} object of class \code{\link{init.net}}.}

\item{control}{An \code{EpiModel} object of class \code{\link{control.net}}.}

\item{s}{Simulation number, used for restarting dependent simulations.}
}
\description{
Sets the parameters, initial conditions, and control settings on
             the data object, necessary for checkpointing simulations.
}
\details{
When running a stochastic network model from checkpointed data, it is not
necessary to run the originally specified initialization module. Instead, the
initialization module should reset the parameters, initial conditions, and
control settings back onto the data object.

This module is intended to be used in the context of running simulations on
high-performance computing settings using \code{\link{netsim_hpc}}. That
function automatically replaces the original initialization function with this
checkpointed version when the simulation is in a checkpoint state.
}

