% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swfcalib_helpers.R
\name{make_calibrated_scenario}
\alias{make_calibrated_scenario}
\title{Make an EpiModel scenario using the result of an `swfcalib` calibration}
\usage{
make_calibrated_scenario(calib_object)
}
\arguments{
\item{calib_object}{a formatted calibration object}
}
\description{
Make an EpiModel scenario using the result of an `swfcalib` calibration
}
