% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netsim_scenarios.R
\name{make_save_elements}
\alias{make_save_elements}
\title{Create the `save_elements` vector for `netsim_run_one_scenario`}
\usage{
make_save_elements(save_pattern)
}
\arguments{
\item{save_pattern}{A character vector of what should be kept in the final
`netsim` objects. It can contain the names of the elements as well as:
"simple" (defautlt) to only keep "epi", "param" and "control"; "restart" to
get the elements required to restart from such file; "all" to not trim the
object at all. `c("simple", "el.cuml")` is an example of a valid pattern to
save "epi", "param", "control" and "el.cuml".}
}
\description{
Helper function to create the `save_elements` character vector according to
the `save_pattern`.
}
