% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netsim_scenarios.R
\name{merge_netsim_scenarios}
\alias{merge_netsim_scenarios}
\title{Create a Single Sim File per Scenarios Using the Files From
\code{netsim_scenarios}}
\usage{
merge_netsim_scenarios(
  sim_dir,
  output_dir,
  keep.transmat = TRUE,
  keep.network = TRUE,
  keep.nwstats = TRUE,
  keep.other = TRUE,
  param.error = FALSE,
  keep.diss.stats = TRUE,
  truncate.at = NULL,
  keep.run = TRUE,
  keep.cumulative.edgelist = FALSE,
  keep.attr.history = TRUE
)
}
\arguments{
\item{sim_dir}{The folder where the simulation files are to be stored.}

\item{output_dir}{The folder where the merged files will be stored.}

\item{keep.transmat}{If \code{TRUE}, keep the transmission matrices from the
original \code{x} and \code{y} elements. Note: transmission matrices
only saved when (\code{save.transmat == TRUE}).}

\item{keep.network}{If \code{TRUE}, keep the \code{networkDynamic} objects
from the original \code{x} and \code{y} elements. Note: network
only saved when (\code{tergmLite == FALSE}).}

\item{keep.nwstats}{If \code{TRUE}, keep the network statistics (as set by
the \code{nwstats.formula} parameter in \code{control.net}) from
the original \code{x} and \code{y} elements.}

\item{keep.other}{If \code{TRUE}, keep the other simulation elements (as set
by the \code{save.other} parameter in \code{control.net}) from the
original \code{x} and \code{y} elements.}

\item{param.error}{If \code{TRUE}, if \code{x} and \code{y} have different
params (in \code{\link[EpiModel:param.net]{param.net()}}) or controls (passed in
\code{\link[EpiModel:control.net]{control.net()}}) an error will prevent the merge. Use
\code{FALSE} to override that check.}

\item{keep.diss.stats}{If \code{TRUE}, keep \code{diss.stats} from the
original \code{x} and \code{y} objects.}

\item{truncate.at}{Time step at which to left-truncate the time series.}

\item{keep.run}{If \code{TRUE}, keep the \code{run} sublists (as set by the
\code{save.run} parameter in \code{control.net}) from the original \code{x} and
\code{y} elements. These are required to use the merged object as a
restart point in \code{\link[EpiModel:netsim]{netsim()}}.}

\item{keep.cumulative.edgelist}{If \code{TRUE}, keep the cumulative edgelists
(as set by the \code{save.cumulative.edgelist} parameter in
\code{control.net}) from the original \code{x} and \code{y} elements. \code{FALSE} by
default, as these grow with the length of the simulation.}

\item{keep.attr.history}{If \code{TRUE}, keep the recorded histories (as set by
\code{\link[EpiModel:record_attr_history]{record_attr_history()}} and \code{\link[EpiModel:record_raw_object]{record_raw_object()}}) from the original
\code{x} and \code{y} elements. This governs both \code{attr.history} and
\code{raw.records}.}
}
\description{
Create a Single Sim File per Scenarios Using the Files From
\code{netsim_scenarios}
}
