% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netsim_scenarios.R
\name{merge_netsim_scenarios_tibble}
\alias{merge_netsim_scenarios_tibble}
\title{Create a Single Sim File per Scenarios Using the Files From
\code{netsim_scenarios}}
\usage{
merge_netsim_scenarios_tibble(
  sim_dir,
  output_dir,
  steps_to_keep,
  cols = dplyr::everything()
)
}
\arguments{
\item{sim_dir}{The folder where the simulation files are to be stored.}

\item{output_dir}{The folder where the merged files will be stored.}

\item{steps_to_keep}{Numbers of time steps add the end of the simulation to
keep in the \code{data.frame}.}

\item{cols}{\if{html}{\out{<tidy-select>}} columns to keep in the \code{data.frame}. By default all
columns are kept. And in any case, the \code{batch_number}, \code{sim} and \code{time}
are always kept.}
}
\description{
Create a Single Sim File per Scenarios Using the Files From
\code{netsim_scenarios}
}
