% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_simfiles.R
\name{merge_simfiles}
\alias{merge_simfiles}
\title{Save Simulation Data from Stochastic Network Models}
\usage{
merge_simfiles(
  simno,
  ftype = "min",
  indir = "data/",
  vars = NULL,
  truncate.at = NULL,
  verbose = TRUE
)
}
\arguments{
\item{simno}{First components of the simulation number in the standard format
written by \code{\link{savesim}} (see details).}

\item{ftype}{Type of file to be merged, with either \code{"min"} for compacted
files or \code{"max"} for large files. File availability depends on
what files were saved in \code{\link{savesim}}.}

\item{indir}{File directory relative to working directory where simulation
files are stored.}

\item{vars}{Character vector of variables stored in \code{epi} sub-list to
retain in output data. If any variables are specified, then network
statistics and other ancillary data are removed.}

\item{truncate.at}{Left-truncates a simulation epidemiological summary
statistics and network statistics at a specified time step.}

\item{verbose}{If \code{TRUE}, print file load progress to console.}
}
\description{
Saves an Rdata file containing stochastic network model output
             from \code{netsim} function calls with time-stamped file names.
}
\details{
This function merges individual simulation runs stored in separate Rdata files
into one larger output object for analysis. This function would typically be
used after running \code{\link{netsim_hpc}} with an array job specification
(see the vignette) in order to combine individual blocks of simulations into
one complete set.

The \code{simno} argument must therefore be specified as the first component
of the simulation number: what would be passed to the \code{-v} parameter
in \code{qsub}. For example, if one would like to aggregate the two files for
simulation number 1 stored in the \code{sim.n1.1.*} and \code{sim.n1.2.*} files,
the \code{simno} argument would be \code{1}.
}
