% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mod_thresh.R
\name{mod_fit}
\alias{mod_fit}
\title{Find Best Fitting Modeling on Time-Series Equilibrium and Epidemic
       Prevalence}
\usage{
mod_fit(dir, job.nos = "all", nsteps = 1000, threshold = 0.001,
  prev = NULL)
}
\arguments{
\item{dir}{File directory in which the simulations are stored.}

\item{job.nos}{Set of simulation job numbers, as an integer, for fit calculations
to be run. If \code{"all"}, then will use all simulation data within the
specified directory.}

\item{nsteps}{For the equilibrium threshold calculation, number of time steps
from the end of the simulation over which to calculate equilbrium,
following the methods in \code{calc_eql} in the EpiModel package.}

\item{threshold}{Threshold level at which the two decision rules will be
made}

\item{prev}{Observed disease prevalence to which simulation prevalence should
       be fit. If \code{NULL}, then only perform the equilibrium decision rule.}
}
\description{
Calculates whether a set of models have met two thresholds for
             dynamic model fit: stable equilibrium prevalence and fit to an
             observed disease prevalence.
}
\examples{
feql <- mod_fit(dir = "data", prev = 0.0475)
}

