% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/netsim_hpc.R
\name{netsim_hpc}
\alias{netsim_hpc}
\title{Stochastic Network Models on High-Performance Computing Systems}
\usage{
netsim_hpc(x, param, init, control, save.min = TRUE, save.max = FALSE,
  compress = FALSE)
}
\arguments{
\item{x}{Character vector containing the file location and name of an Rdata
file where an object of class \code{netest} is stored. Alternatively,
if restarting a previous simulation, this may be a file name for an
object of class \code{netsim}.}

\item{param}{Model parameters, as an object of class \code{param.net}.}

\item{init}{Initial conditions, as an object of class \code{init.net}.}

\item{control}{Control settings, as an object of class \code{control.net}.}

\item{save.min}{Argument passed to \code{\link{savesim}}.}

\item{save.max}{Argument passed to \code{\link{savesim}}.}

\item{compress}{Matches the \code{compress} argument for the \code{\link{save}}
       function.}
}
\description{
Simulates stochastic network epidemic models for infectious
             disease in parallel.
}
\details{
This function provides a systematic method to running stochastic network
models in parallel on high-performance computing systems. The function wraps
\code{\link{netsim_par}} that establishes the parallelization of the
underlying sequential simulations run in \code{netsim}.

The main purpose of using \code{netsim_hpc} is for a standardized checkpointing
method. Checkpointing is defined as incrementally saving simulation data for
the purpose of reloading it if a simulation job is canceled and restarted. If
checkpointing is not needed, users are advised to run their models directly
with the \code{\link{netsim_par}} function.

This function performs the following tasks:
\enumerate{
  \item Check for the existence of checkpointed data, using the
        \code{\link{check_cp}} function. If CP data are available, a
        checkpointed model will be run, else a new model will be run.
  \item Create a checkpoint directory if one does not exist at
        "data/sim<simno>".
  \item Set a save interval of 100 time steps if one does not already exist
        on the control settings.
  \item Resets the initialize module function to \code{\link{initialize_cp}}
        if in checkpoint state.
  \item Run the simulation, either new or checkpointed, with a call to
        \code{\link{netsim_par}}.
  \item Save the completed simulation data, using the functionality of
        \code{\link{savesim}}.
  \item Remove any files in the "verb/" subdirectory, which is typically
        used to store incremental model tracking text files.
  \item Remove the checkpointed data and file directory created in step 1.
}

Note that the \code{x} argument must specify a \strong{file name} in a
character string, rather than a \code{netest} or \code{netsim} class object
directly. This is mainly for efficency purposes in running the models in
parallel.
}

