% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mod_fit.R
\name{print_modfit}
\alias{print_modfit}
\title{Printing Method for mod_fit Objects}
\usage{
print_modfit(x, out = "all", print.gap = 3)
}
\arguments{
\item{x}{Output object from \code{\link{mod_fit}} function.}

\item{out}{Either \code{"all"} for full table, \code{"eql"} for table limited
to those jobs passing the equilibrium test, or \code{"epi"} for table
limited to those jobs passing both the equilibrium and epidemiology
test.}

\item{print.gap}{Spacing method passed to \code{print} generic.}
}
\description{
Prints output from \code{mod_fit} calls.
}

