% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub_master.R
\name{qsub_master}
\alias{qsub_master}
\title{Create qsub Bash Shell Script Parameter Combination}
\usage{
qsub_master(
  outfile = "master.sh",
  runsimfile = "runsim.sh",
  simno.start = 1,
  nsubjobs = 4,
  backfill = TRUE,
  email = FALSE,
  append = FALSE,
  vars
)
}
\arguments{
\item{outfile}{Name of the output bash shell script file. If \code{""}, then
will print to console.}

\item{runsimfile}{Name of the bash shell script file that contains the R batch
commandsis}

\item{simno.start}{Starting number for the \code{SIMNO} variable. If set to
\code{"auto"} and \code{append=TRUE}, will read the lines of \code{outfile}
and start numbering at one after the previous maximum.}

\item{nsubjobs}{Number of sub/array jobs to run per simulation set.}

\item{backfill}{If \code{TRUE}, use the backfill queue to submit jobs. If
numeric, will specify the first X jobs on the grid as non-backfill.}

\item{email}{If \code{TRUE}, send email on job termination or completion.}

\item{append}{If \code{TRUE}, will append lines to a previously created shell
script.}

\item{vars}{A list of parameters with varying values (see example).}
}
\description{
Creates a master-level qsub script given a set of parameter
             combinations implied by one or more parameters passed in.
}
\examples{
vars <- list(A = 1:10, B = seq(0.5, 1.5, 0.5))
qsub_master(vars = vars, outfile = "")

qsub_master(vars = vars, outfile = "", backfill = TRUE)
qsub_master(vars = vars, outfile = "", backfill = 10)

}
