% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/save_cpdata.R
\name{save_cpdata}
\alias{save_cpdata}
\title{Saves for Network Simulation Rdata Files for Checkpointing}
\usage{
save_cpdata(dat, at)
}
\arguments{
\item{dat}{A master data object used in models simulated with \code{netsim}.}

\item{at}{Current time step}
}
\description{
Module to save simulation data from stochastic network models
             to disk at specified time intervals.
}
\details{
This module saves data to a standardized location with standardized file names
for the purposes of checkpointing. This is intended to be used when running
these simulations with \code{\link{netsim_hpc}}, and will be automatically
inserted into the workflow when this is done.
}

