% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/savesim.R
\name{savesim}
\alias{savesim}
\title{Save Simulation Data from Stochastic Network Models}
\usage{
savesim(sim, dataf = TRUE, save.min = TRUE, save.max = TRUE,
  compress = "xz")
}
\arguments{
\item{sim}{An \code{EpiModel} object of class \code{netsim} to be saved to an
Rdata file.}

\item{dataf}{If \code{TRUE}, saves data file within a subfolder called "data"
relative to the current working directory, otherwise saves file to
current working directory. A data folder will be created if one does
not already exist.}

\item{save.min}{If \code{TRUE}, saves a small version of the \code{netsim}
object in which large elements of the data structure like the network
object and the transmission data frame are removed. The resulting
name for this small file will have ".min" appended at the end.}

\item{save.max}{If \code{TRUE}, saves the full \code{netsim} object without
any deletions.}

\item{compress}{Matches the \code{compress} argument for the \code{\link{save}}
       function. Set to \code{FALSE} for no compression.}
}
\description{
Saves an Rdata file containing stochastic network model output
             from \code{netsim} function calls with time-stamped file names.
}
\details{
This function provides an automated method for saving a time-stamped Rdata
file containing the simulation number of a stochastic network model run
with \code{netsim}. This is used by default by the \code{\link{netsim_hpc}}
function.
}

