% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbatch_master.R
\name{sbatch_master}
\alias{sbatch_master}
\title{Create sbatch Bash Shell Script with Parameter Combination}
\usage{
sbatch_master(vars, outfile = "", runsimfile = "runsim.sh",
  simno.start, narray = 0, ckpt = FALSE, append = FALSE)
}
\arguments{
\item{vars}{A list of parameters with varying values (see examples below).}

\item{outfile}{Name of the output bash shell script file to write. If \code{""}, 
then will print to console.}

\item{runsimfile}{Name of the bash shell script file that contains the R batch
commands}

\item{simno.start}{Starting number for the \code{SIMNO} variable. If missing
and \code{append=TRUE}, will read the lines of \code{outfile}
and start numbering at one after the previous maximum.}

\item{narray}{Number of array jobs to run per scenario set.}

\item{ckpt}{If \code{TRUE}, use the checkpoint queue to submit jobs. If
numeric, will specify the first X jobs on the grid as non-backfill.}

\item{append}{If \code{TRUE}, will append lines to a previously created shell
script. New simno will either start with value of \code{simno.start}
or the previous value if missing.}
}
\description{
Creates a master-level SLURM::sbatch script given a set of parameter
             combinations implied by environmental arguments used as parameters.
}
\examples{
vars <- list(A = 1:10, B = seq(0.5, 1.5, 0.5))
sbatch_master(vars = vars)
sbatch_master(vars = vars, narray = 10)
sbatch_master(vars = vars, narray = 4, ckpt = TRUE)
sbatch_master(vars = vars, narray = 4, ckpt = 10)
sbatch_master(vars = vars, simno.start = 1000)

\dontrun{
sbatch_master(vars = vars, narray = 4, simno.start = 1000, outfile = "master.sh")
sbatch_master(vars = vars, narray = 4, append = TRUE, outfile = "master.sh")

sbatch_master(vars = vars, simno.start = 1000, outfile = "master.sh")
sbatch_master(vars = vars, simno.start = 2000, append = TRUE, outfile = "master.sh")
}

}
