% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netsim_scenarios.R
\name{step_tmpl_netsim_scenarios}
\alias{step_tmpl_netsim_scenarios}
\title{Step template to run EpiModel network simulations with scenarios}
\usage{
step_tmpl_netsim_scenarios(
  path_to_x,
  param,
  init,
  control,
  scenarios_list,
  n_rep,
  n_cores,
  output_dir,
  libraries = NULL,
  setup_lines = NULL,
  max_array_size = NULL,
  ...
)
}
\arguments{
\item{path_to_x}{Path to a fitted network model or a restart point saved
with \code{saveRDS} (See the \code{x} argument to the \code{EpiModel::netsim}
function), or to a restart pool directory (see \code{make_restart_pool}).}

\item{param}{Model parameters, as an object of class \code{\link[EpiModel]{param.net}}.
Includes transmission probability (\code{inf.prob}), act rate
(\code{act.rate}), recovery rate (\code{rec.rate}), and demographic rates
for models with vital dynamics. Custom parameters for extended
models may also be passed through \code{param.net}.}

\item{init}{Initial conditions, as an object of class \code{\link[EpiModel]{init.net}}.
Specifies the initial number of infected nodes (\code{i.num}) and,
for SIR models, recovered nodes (\code{r.num}). For two-group models,
the corresponding \code{.g2} parameters are also required.}

\item{control}{Control settings, as an object of class \code{\link[EpiModel]{control.net}}.
Key settings include \code{type} (disease model: \code{"SI"}, \code{"SIR"}, or
\code{"SIS"}), \code{nsteps} (number of time steps), \code{nsims} (number of
simulations), \code{tergmLite} (lightweight mode for performance), and
\code{resimulate.network} (required for models with vital dynamics).
For extended models, custom module functions are also passed here.}

\item{scenarios_list}{A list of scenarios to be run. Produced by the
\code{EpiModel::create_scenario_list} function}

\item{n_rep}{The number of replication to be run for each scenario.}

\item{n_cores}{The number of CPUs on which the simulations will be run.}

\item{output_dir}{The folder where the simulation files are to be stored.}

\item{libraries}{A character vector containing the name of the libraries
required for the model to run. (e.g. EpiModelHIV or EpiModelCOVID)}

\item{setup_lines}{(optional) a vector of bash lines to be run first.
This can be used to load the required modules (like R, python, etc).}

\item{max_array_size}{maximum number of array jobs to be submitted at once.
It should be strictly less than the maximum number of jobs you are allowed
to submit to \code{slurm} on your HPC. If the number of jobs is greater than
\code{max_array_size}, they will be submitted in multiple batches. The
submission of the next batch is automatically handled by \code{slurmworkflow}.}

\item{...}{arguments to vectorize over (vectors or lists of strictly
positive length, or all of zero length).  See also ‘Details’.}
}
\value{
a template function to be used by \code{add_workflow_step}
}
\description{
This step template is similar to \code{netsim_scenarios} but for the HPC. It uses
\code{slurmworkflow::step_tmpl_map} internally and should be used as any
\code{slurmworkflow} step. For details, see \code{netsim_scenarios} documentation.
The simulations of a batch are run in parallel, one \code{netsim} call per
simulation (see \code{netsim_path_wrapper}), by default with
\code{future::plan("multicore", workers = n_cores)}.
Using \verb{control$future.use.plan <- future::tweak(<your plan>)} will bypass
this setting.
}
\section{Checkpointing}{

This function takes care of editing \code{.checkpoint.dir} to create unique sub
directories for each scenario, batch and simulation:
"<.checkpoint.dir>/sim__\if{html}{\out{<scenario id>}}__<batch_num>/sim_\if{html}{\out{<k>}}". The
\code{EpiModel::control.net} way of setting up checkpoints can be used
transparently.
}

\section{Step Template}{


Step Templates are helper functions to be used within \code{add_workflow_step}.
Some basic ones are provided by the \code{slurmworkflow} package. They instruct
the workflow to run either a bash script, a set of bash lines given as a
character vector or an R script.
Additional Step Templates can be created to simplify specific tasks. The
easiest way to do so is as wrappers around existing templates.

}

