% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurmworkflow_helpers.R
\name{step_tmpl_renv_restore}
\alias{step_tmpl_renv_restore}
\title{Step template to update a project \code{renv}}
\usage{
step_tmpl_renv_restore(git_branch, setup_lines = NULL)
}
\arguments{
\item{git_branch}{The git branch that the project is supposed to follow. If
the project is not following the right branch, this step will error.}

\item{setup_lines}{(optional) a vector of bash lines to be run first.
This can be used to load the required modules (like R, python, etc).}
}
\value{
a template function to be used by \code{add_workflow_step}
}
\description{
This template makes the step run `git pull` and \code{renv::restore()}. This
could help ensure that the project is up to date when running the rest of the
workflow.
See \code{slurmworkflow::step_tmpl_bash_lines} for details on step templates
}
