% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurmworkflow_helpers.R
\name{swf_configs_hyak}
\alias{swf_configs_hyak}
\title{Preset of Configuration for the HYAK Cluster}
\usage{
swf_configs_hyak(hpc = "klone", partition = "csde")
}
\arguments{
\item{hpc}{Which HPC to use on HYAK (either "klone" or "mox")}

\item{partition}{Which partition to use on HYAK (either "csde" or "ckpt")}
}
\value{
a list containing `default_sbatch_opts`, `renv_sbatch_opts` and
   `r_loader` (see the "hpc_configs" section)
}
\description{
Preset of Configuration for the HYAK Cluster
}
\section{hpc_configs}{

- `default_sbatch_opts` is a list of sbatch options to be passed to
`slurmworkflow::create_workflow`.
- `renv_sbatch_opts` is a list of sbatch options to be passed to
`slurmworkflow::step_tmpl_renv_restore`. It provides sane defaults for
building the dependencies of an R project using `renv`
- `r_loader` is a set of bash lines to make the R software available. This is
passed to the `setup_lines` arguments of the `slurmworkflow::step_tmpl_`
functions that requires it.
}

