% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swfcalib_helpers.R
\name{swfcalib_proposal_to_scenario}
\alias{swfcalib_proposal_to_scenario}
\title{Convert an swfcalib Proposal into an EpiModel Scenario}
\usage{
swfcalib_proposal_to_scenario(proposal)
}
\arguments{
\item{proposal}{an swfcalib formatted proposal}
}
\value{
an EpiModel Scenario
}
\description{
Convert an swfcalib Proposal into an EpiModel Scenario
}
