% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModelHPC-package.r
\docType{package}
\name{EpiModelHPC-package}
\alias{EpiModelHPC-package}
\alias{EpiModelHPC}
\title{EpiModel Extensions for High-Performance Computing}
\description{
Extension package to EpiModel for running batched scenario simulations on Slurm-based high-performance computing systems. Provides scenario orchestration with 'netsim_scenarios()', 'slurmworkflow' step templates for HPC job pipelines, 'swfcalib' calibration helpers, and utilities for merging and post-processing simulation output files.
}
\details{
Extension package to EpiModel for running batched scenario simulations on
Slurm-based high-performance computing systems. Provides scenario
orchestration, slurmworkflow step templates, swfcalib calibration helpers,
and utilities for merging and post-processing simulation output files.
}
\references{
The main website for EpiModel is at \url{http://epimodel.org/}.
The source code is hosted on GitHub at
\url{http://github.com/EpiModel/EpiModelHPC}. Bug reports and
feature requests may be filed there.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://epimodel.org/}
  \item \url{http://epimodel.github.io/EpiModelHPC/}
  \item Report bugs at \url{https://github.com/EpiModel/EpiModelHPC/issues}
}

}
\author{
\strong{Maintainer}: Samuel Jenness \email{samuel.m.jenness@emory.edu}

}
\keyword{package}
