% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netsim_scenarios.R
\name{get_scenarios_name_parts}
\alias{get_scenarios_name_parts}
\title{Helper function to access the file name elements of scenarios}
\usage{
get_scenarios_name_parts(scenario_dir)
}
\arguments{
\item{scenario_dir}{the directory where `netsim_scenarios` saved it's
simulations.}
}
\value{
a `list` with three elements: `file_name_list`, a character vector of
the full paths of the simulation files; `scenario_name_list`, a character
vector of the associated scenario name; `batch_number_list`, a character
vector of the associated batch number.
}
\description{
This function returns the list of simulation files and the corresponding
scenario name and batch number present in a given directory. It is meant to
be used after `netsim_scenarios` or `step_tmpl_netsim_scenarios`.
}
