% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netsim_scenarios.R
\name{netsim_run_one_scenario}
\alias{netsim_run_one_scenario}
\title{Inner function called by `netsim_scenarios` and `step_tmpl_netsim_scenarios`}
\usage{
netsim_run_one_scenario(
  scenario,
  batch_num,
  est,
  param,
  init,
  control,
  libraries,
  output_dir,
  n_batch,
  n_rep,
  n_cores
)
}
\description{
Inner function called by `netsim_scenarios` and `step_tmpl_netsim_scenarios`
}
