% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swfcalib_helpers.R
\name{netsim_run_swfcalib_scenario}
\alias{netsim_run_swfcalib_scenario}
\title{Run one `netsim` call with the result of an `swfcalib` calibration}
\usage{
netsim_run_swfcalib_scenario(
  calib_object,
  batch_num,
  x,
  param,
  init,
  control,
  libraries,
  output_dir,
  n_batch,
  n_rep,
  n_cores,
  save_all,
  save_elements
)
}
\arguments{
\item{calib_object}{a formatted calibration object}

\item{batch_num}{The batch number, calculated from the number of replications
and CPUs required.}

\item{x}{Fitted network model object, as an object of class \code{netest}.
Alternatively, if restarting a previous simulation, may be an object
of class \code{netsim}.}

\item{param}{Model parameters, as an object of class \code{param.net}.}

\item{init}{Initial conditions, as an object of class \code{init.net}.}

\item{control}{Control settings, as an object of class
\code{control.net}.}

\item{libraries}{A character vector containing the name of the libraries
required for the model to run. (e.g. EpiModelHIV or EpiModelCOVID)}

\item{output_dir}{The folder where the simulation files are to be stored.}

\item{n_batch}{The number of batches to be run `ceiling(n_rep / n_cores)`.}

\item{n_rep}{The number of replication to be run for each scenario.}

\item{n_cores}{number of cores to run the processing on}

\item{save_all}{A flag instructing to save the result of the
`EpiModel::netsim` call as is if TRUE.}

\item{save_elements}{A character vector of elements to keep from the
`netsim` object if `save_all` is `FALSE`}
}
\description{
Run one `netsim` call with the result of an `swfcalib` calibration
}
