% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netsim_scenarios.R
\name{netsim_scenarios_setup}
\alias{netsim_scenarios_setup}
\title{Helper function to configure `netsim_run_one_scenario`}
\usage{
netsim_scenarios_setup(
  est,
  param,
  init,
  control,
  scenarios_list,
  n_rep,
  n_cores,
  output_dir,
  libraries
)
}
\description{
Helper function to configure `netsim_run_one_scenario`
}
