% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netsim_scenarios.R
\name{netsim_scenarios_setup}
\alias{netsim_scenarios_setup}
\title{Helper function to  create the parameters for \code{netsim_run_one_scenario}}
\usage{
netsim_scenarios_setup(
  path_to_x,
  param,
  init,
  control,
  scenarios_list,
  n_rep,
  n_cores,
  output_dir,
  libraries
)
}
\arguments{
\item{path_to_x}{Path to a Fitted network model object saved with \code{saveRDS}.
(See the \code{x} argument to the \code{EpiModel::netsim} function)}

\item{param}{Model parameters, as an object of class \code{param.net}.}

\item{init}{Initial conditions, as an object of class \code{init.net}.}

\item{control}{Control settings, as an object of class
\code{control.net}.}

\item{scenarios_list}{A list of scenarios to be run. Produced by the
\code{EpiModel::create_scenario_list} function}

\item{n_rep}{The number of replication to be run for each scenario.}

\item{n_cores}{The number of CPUs on which the simulations will be run.}

\item{output_dir}{The folder where the simulation files are to be stored.}

\item{libraries}{A character vector containing the name of the libraries
required for the model to run. (e.g. EpiModelHIV or EpiModelCOVID)}
}
\value{
a list of arguments for \code{netsim_run_one_scenario}
}
\description{
Helper function to  create the parameters for \code{netsim_run_one_scenario}
}
