% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swfcalib_helpers.R
\name{netsim_swfcalib_output_setup}
\alias{netsim_swfcalib_output_setup}
\title{Helper function to  create the parameters for `netsim_run_swfcalib_scenario`}
\usage{
netsim_swfcalib_output_setup(
  x,
  param,
  init,
  control,
  calib_object,
  n_rep,
  n_cores,
  output_dir,
  libraries,
  save_pattern
)
}
\arguments{
\item{x}{Fitted network model object, as an object of class \code{netest}.
Alternatively, if restarting a previous simulation, may be an object
of class \code{netsim}.}

\item{param}{Model parameters, as an object of class \code{param.net}.}

\item{init}{Initial conditions, as an object of class \code{init.net}.}

\item{control}{Control settings, as an object of class
\code{control.net}.}

\item{calib_object}{a formatted calibration object}

\item{n_rep}{The number of replication to be run for each scenario.}

\item{n_cores}{number of cores to run the processing on}

\item{output_dir}{The folder where the simulation files are to be stored.}

\item{libraries}{A character vector containing the name of the libraries
required for the model to run. (e.g. EpiModelHIV or EpiModelCOVID)}

\item{save_pattern}{A character vector of what should be kept in the final
`netsim` objects. It can contain the names of the elements as well as:
"simple" (defautlt) to only keep "epi", "param" and "control"; "restart" to
get the elements required to restart from such file; "all" to not trim the
object at all. `c("simple", "el.cuml")` is an example of a valid pattern to
save "epi", "param", "control" and "el.cuml".}
}
\description{
Helper function to  create the parameters for `netsim_run_swfcalib_scenario`
}
