% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merge_simfiles.R
\name{process_simfiles}
\alias{process_simfiles}
\title{Process sub-job simulation files saved as a series of Rdata files.}
\usage{
process_simfiles(simno = NA, indir = "data/", outdir = "data/",
  delete.sub = TRUE)
}
\arguments{
\item{simno}{Simulation number to process.}

\item{indir}{File directory relative to working directory where simulation
files are stored.}

\item{outdir}{File directory relative to working directory where simulation
files should be saved.}

\item{delete.sub}{Delete sub-job files after merge and saving.}
}
\description{
Wraps the \code{merge_simfiles} function to merge all sub-job
             Rdata files and saves into a single output file, with the option
             to delete the sub-job files.
}

