% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_simfiles.R
\name{process_simfiles}
\alias{process_simfiles}
\title{Process sub-job simulation files saved as a series of Rdata files.}
\usage{
process_simfiles(
  simno = NA,
  indir = "data/",
  outdir = "data/",
  vars = NULL,
  truncate.at = NULL,
  min.n,
  nsims,
  compress = "xz",
  delete.sub = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{simno}{Simulation number to process.}

\item{indir}{File directory relative to working directory where simulation
files are stored.}

\item{outdir}{File directory relative to working directory where simulation
files should be saved.}

\item{vars}{Argument passed to \code{\link{merge_simfiles}}.}

\item{truncate.at}{Left-truncates a simulation epidemiological summary
statistics and network statistics at a specified time step.}

\item{min.n}{Integer value for the minimum number of simulation files to be
eligible for processing.}

\item{nsims}{Total number of simulations across all sub-jobs.}

\item{compress}{Argument passed to \code{\link{save}}.}

\item{delete.sub}{Delete sub-job files after merge and saving.}

\item{verbose}{Logical, print progress to console.}
}
\description{
Wraps the \code{merge_simfiles} function to merge all sub-job
             Rdata files and saves into a single output file, with the option
             to delete the sub-job files.
}
