% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_env_vars.R
\name{pull_env_vars}
\alias{pull_env_vars}
\title{Pull Standard Environmental Variables in Slurm Jobs}
\usage{
pull_env_vars()
}
\description{
Pulls four environmental variables commonly used in Slurm jobs
             directly into the Global Environment of an R Script.
}
\examples{
Sys.setenv("SIMNO"=1000)
Sys.setenv("SLURM_ARRAY_TASK_ID"=4)
Sys.setenv("SLURM_NTASKS_PER_NODE"=4)
Sys.setenv("SLURM_ARRAY_TASK_MAX"=10)

ls()
pull_env_vars()
ls()

}
