% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbatch_master.R
\name{sbatch_master}
\alias{sbatch_master}
\title{Create sbatch Bash Shell Script with Parameter Combination}
\usage{
sbatch_master(
  vars,
  expand.vars = TRUE,
  working.dir = "",
  master.file = "",
  runsim.file = "runsim.sh",
  build.runsim = FALSE,
  env.file = "~/loadR.sh",
  rscript.file = "sim.R",
  param.file = NULL,
  param.tag = NULL,
  simno.start,
  nsims = 100,
  ncores = 16,
  narray = NULL,
  ckpt = FALSE,
  append = FALSE,
  mem = "55G",
  walltime = "1:00:00",
  jobname,
  partition.main = "csde",
  partition.ckpt = "ckpt",
  account.main = "csde",
  account.ckpt = "csde-ckpt"
)
}
\arguments{
\item{vars}{A list of parameters with varying values (see examples below).}

\item{expand.vars}{If \code{TRUE}, expand the grid on the individual vars, else
the individual vars must be vectors of equal length.}

\item{working.dir}{Path to write out the \code{master.file}, and if specified
the \code{runsim.file} and \code{param.file}.}

\item{master.file}{Name of the output bash shell script file to write. If 
\code{""}, then will print to console.}

\item{runsim.file}{Name of the bash shell script file that contains the R batch
commands to be executed by \code{sbatch}.}

\item{build.runsim}{If \code{TRUE}, will write out a bash shell script with the
file name \code{runsim.file} that loads the R environment listed in
\code{env.file} and execute \code{Rscript} on the file listed in 
\code{rscript.file}.}

\item{env.file}{Bash shell script to load the R environment desired. Optionally
kept in a user's home directory with the default file name. Example 
script below.}

\item{rscript.file}{Name of the \code{.R} file that contains the primary 
simulation to be executed by \code{Rscript}.}

\item{param.file}{Name of a csv file to write out the list of varying 
parameters and simulation numbers set within the function.}

\item{param.tag}{Character string for current scenario batch added to 
param.sheet.}

\item{simno.start}{Starting number for the \code{SIMNO} variable. If missing
and \code{append=TRUE}, will read the lines of \code{outfile}
and start numbering at one after the previous maximum.}

\item{nsims}{Total number of simulations across all array jobs.}

\item{ncores}{Number of cores per node to use within each Slurm job.}

\item{narray}{Number of array batches within each Slurm job. If `NULL`, then
will use `nsims/ncores` array batches.}

\item{ckpt}{If \code{TRUE}, use the checkpoint queue to submit jobs. If
numeric, will specify the first X jobs on the grid as non-backfill.}

\item{append}{If \code{TRUE}, will append lines to a previously created shell
script. New simno will either start with value of \code{simno.start}
or the previous value if missing.}

\item{mem}{Amount of memory needed per node within each Slurm job.}

\item{walltime}{Amount of clock time needed per Slurm job.}

\item{jobname}{Job name assigned to Slurm job. If unspecified, defaults to the
simulation number in each job.}

\item{partition.main}{Name of primary HPC partition (passed to -p).}

\item{partition.ckpt}{Name of checkpoint HPC partition (passed to -p).}

\item{account.main}{Name of primary account (passed to -A).}

\item{account.ckpt}{Name of checkpoint account (passed to -A).}
}
\description{
Creates a master-level SLURM::sbatch script given a set of parameter
             combinations implied by environmental arguments used as parameters.
}
\examples{
# Examples printing to console
vars <- list(A = 1:5, B = seq(0.5, 1.5, 0.5))
sbatch_master(vars)
sbatch_master(vars, nsims = 250)
sbatch_master(vars, ckpt = TRUE)
sbatch_master(vars, nsims = 50, ckpt = 10)
sbatch_master(vars, simno.start = 1000)
sbatch_master(vars, jobname = "epiSim")

\dontrun{
# Full-scale example writing out files
sbatch_master(vars, nsims = 50, simno.start = 1000, build.runsim = TRUE,
              master.file = "master.sh", param.sheet = "params.csv")
sbatch_master(vars, nsims = 50, append = TRUE, 
              master.file = "master.sh", param.sheet = "params.csv")

## Example bash environment file
#!/bin/bash

. /gscratch/csde/sjenness/spack/share/spack/setup-env.sh
module load gcc-8.2.0-gcc-4.8.5-rhsxipz
module load r-3.5.1-gcc-8.2.0-4suigve

}

}
