% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netsim_scenarios.R
\name{step_tmpl_merge_netsim_scenarios}
\alias{step_tmpl_merge_netsim_scenarios}
\title{Step Template to Create a Single Sim File per Scenarios Using the Files From
\code{netsim_scenarios}}
\usage{
step_tmpl_merge_netsim_scenarios(
  sim_dir,
  output_dir,
  keep.transmat = TRUE,
  keep.network = TRUE,
  keep.nwstats = TRUE,
  keep.other = TRUE,
  param.error = FALSE,
  keep.diss.stats = TRUE,
  truncate.at = NULL,
  n_cores = 1,
  setup_lines = NULL
)
}
\arguments{
\item{sim_dir}{The folder where the simulation files are to be stored.}

\item{output_dir}{The folder where the merged files will be stored.}

\item{keep.transmat}{If \code{TRUE}, keep the transmission matrices from the
original \code{x} and \code{y} elements. Note: transmission matrices
only saved when (\code{save.transmat == TRUE}).}

\item{keep.network}{If \code{TRUE}, keep the \code{networkDynamic} objects
from the original \code{x} and \code{y} elements. Note: network
only saved when (\code{tergmLite == FALSE}).}

\item{keep.nwstats}{If \code{TRUE}, keep the network statistics (as set by
the \code{nwstats.formula} parameter in \code{control.netsim}) from
the original \code{x} and \code{y} elements.}

\item{keep.other}{If \code{TRUE}, keep the other simulation elements (as set
by the \code{save.other} parameter in \code{control.netsim}) from the
original \code{x} and \code{y} elements.}

\item{param.error}{If \code{TRUE}, if \code{x} and \code{y} have different
params (in \code{\link[EpiModel]{param.net}}) or controls (passed in
\code{\link[EpiModel]{control.net}}) an error will prevent the merge. Use
\code{FALSE} to override that check.}

\item{keep.diss.stats}{If \code{TRUE}, keep \code{diss.stats} from the
original \code{x} and \code{y} objects.}

\item{n_cores}{Parallelize the process over \code{n_cores} (default = 1)}

\item{setup_lines}{(optional) a vector of bash lines to be run first.
This can be used to load the required modules (like R, python, etc).}
}
\value{
a template function to be used by \code{add_workflow_step}
}
\description{
Step Template to Create a Single Sim File per Scenarios Using the Files From
\code{netsim_scenarios}
}
\section{Step Template}{


Step Templates are helper functions to be used within \code{add_workflow_step}.
Some basic ones are provided by the \code{slurmworkflow} package. They instruct
the workflow to run either a bash script, a set of bash lines given as a
character vector or an R script.
Additional Step Templates can be created to simplify specific tasks. The
easiest way to do so is as wrappers around existing templates.

}

