% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netsim_scenarios.R
\name{step_tmpl_merge_netsim_scenarios_tibble}
\alias{step_tmpl_merge_netsim_scenarios_tibble}
\title{Step Template to Create a Single Sim File per Scenarios Using the Files From
\code{netsim_scenarios}}
\usage{
step_tmpl_merge_netsim_scenarios_tibble(
  sim_dir,
  output_dir,
  steps_to_keep,
  cols = dplyr::everything(),
  n_cores = 1,
  setup_lines = NULL
)
}
\arguments{
\item{sim_dir}{The folder where the simulation files are to be stored.}

\item{output_dir}{The folder where the merged files will be stored.}

\item{setup_lines}{(optional) a vector of bash lines to be run first.
This can be used to load the required modules (like R, python, etc).}
}
\value{
a template function to be used by \code{add_workflow_step}
}
\description{
Step Template to Create a Single Sim File per Scenarios Using the Files From
\code{netsim_scenarios}
}
\section{Step Template}{


Step Templates are helper functions to be used within \code{add_workflow_step}.
Some basic ones are provided by the \code{slurmworkflow} package. They instruct
the workflow to run either a bash script, a set of bash lines given as a
character vector or an R script.
Additional Step Templates can be created to simplify specific tasks. The
easiest way to do so is as wrappers around existing templates.

}

