% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurmworkflow_helpers.R
\name{swf_configs_rsph}
\alias{swf_configs_rsph}
\title{Preset of Configuration for the RSPH Cluster}
\usage{
swf_configs_rsph(
  partition = "preemptable",
  r_version = "4.5.1",
  mail_user = NULL,
  cleanup_workers = TRUE
)
}
\arguments{
\item{partition}{Which partition to use on RSPH (either "compute" or
"epimodel")}

\item{r_version}{Which version of R to load}

\item{mail_user}{The mail address to send the messages to, default is NULL
(see 'sbatch --mail-type' argument)}

\item{cleanup_workers}{If \code{TRUE} (the default), append
\code{\link{swf_cleanup_r_workers}} to \code{r_loader} so each step reaps
its own R workers on cancel or timeout. RSPH has no cgroup containment, so
without this, workers survive \code{scancel} and squat on cores SLURM has
re-allocated, starving later tasks. See \code{\link{swf_cleanup_r_workers}}.}
}
\value{
a list containing \code{default_sbatch_opts}, \code{renv_sbatch_opts}
and \code{r_loader} (see the "hpc_configs" section)
}
\description{
Preset of Configuration for the RSPH Cluster
}
\section{Node setup}{

\code{r_loader} sources the shared spack environment, unloads everything, and
loads R. It deliberately does not load git: git is on the default path on
RSPH, and the project template has not loaded it for some time. The
\code{git_version} argument was removed accordingly.
}

\section{hpc_configs}{

\enumerate{
\item \code{default_sbatch_opts} is a list of sbatch options to be passed to
\code{slurmworkflow::create_workflow}.
\item \code{renv_sbatch_opts} is a list of sbatch options to be passed to
\code{slurmworkflow::step_tmpl_renv_restore}. It provides sane defaults for
building the dependencies of an R project using \code{renv}
\item \code{r_loader} is a set of bash lines to make the R software available.
This is passed to the \code{setup_lines} arguments of the
\code{slurmworkflow::step_tmpl_} functions that requires it.
}
}

