% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurmworkflow_helpers.R
\name{swf_configs_rsph}
\alias{swf_configs_rsph}
\title{Preset of Configuration for the RSPH Cluster}
\usage{
swf_configs_rsph(partition = "preemptable")
}
\arguments{
\item{partition}{Which partition to use on RSPH (either "compute" or
"epimodel")}
}
\value{
a list containing `default_sbatch_opts`, `renv_sbatch_opts` and
   `r_loader` (see the "hpc_configs" section)
}
\description{
Preset of Configuration for the RSPH Cluster
}
\section{hpc_configs}{

- `default_sbatch_opts` is a list of sbatch options to be passed to
`slurmworkflow::create_workflow`.
- `renv_sbatch_opts` is a list of sbatch options to be passed to
`slurmworkflow::step_tmpl_renv_restore`. It provides sane defaults for
building the dependencies of an R project using `renv`
- `r_loader` is a set of bash lines to make the R software available. This is
passed to the `setup_lines` arguments of the `slurmworkflow::step_tmpl_`
functions that requires it.
}

