% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swfcalib_helpers.R
\name{swfcalib_proposal_to_scenario}
\alias{swfcalib_proposal_to_scenario}
\title{Convert an swfcalib Proposal into an EpiModel Scenario}
\usage{
swfcalib_proposal_to_scenario(proposal, id = NULL)
}
\arguments{
\item{proposal}{an swfcalib formatted proposal}

\item{id}{the \code{.scenario.id} for the scenario. If \code{NULL}, use the
\code{.proposal_index} or "default" if the former is \code{NULL} as well.}
}
\value{
an EpiModel scenario
}
\description{
Convert an swfcalib Proposal into an EpiModel Scenario
}
